Make shortform collapsible (and start adding it to the frontpage)
